requirements.txt
ansicon==1.89.0
audioop-lts==0.2.2
awscli==1.42.30
awsebcli==3.25.1
bcrypt==5.0.0
blessed==1.22.0
blinker==1.9.0
boto3==1.40.30
botocore==1.40.30
cement==2.10.14
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
clr_loader==0.2.7.post0
cobble==0.1.4
colorama==0.4.6
cryptography==46.0.2
decorator==5.2.1
Deprecated==1.2.18
docutils==0.19
fabric==3.2.2
Flask==3.1.2
gunicorn==23.0.0
idna==3.10
invoke==2.2.0
itsdangerous==2.2.0
Jinja2==3.1.6
jinxed==1.3.0
jmespath==1.0.1
lxml==6.0.1
mammoth==1.11.0
MarkupSafe==3.0.2
packaging==24.2
paramiko==4.0.0
pathspec==0.12.1
pyasn1==0.6.1
PyAudio==0.2.14
pycparser==2.23
pydub==0.25.1
PyNaCl==1.6.0
PyPDF2==3.0.1
pypiwin32==223
python-dateutil==2.9.0.post0
python-docx==1.2.0
pythonnet==3.0.5
pywin32==311
PyYAML==6.0.2
requests==2.32.5
rsa==4.7.2
s3transfer==0.14.0
semantic-version==2.10.0
setuptools==80.9.0
six==1.17.0
termcolor==2.5.0
typing_extensions==4.15.0
urllib3==1.26.20
wcwidth==0.2.14
Werkzeug==3.1.3
wrapt==1.17.3
